+++ Program.cs
@@ -6,6 +6,15 @@ namespace wow.tools.local
     {
         public static void Main(string[] args)
         {
+            // As the absolute first thing, ensure we're actually running in the correct dir
+            if (!Directory.Exists("wwwroot"))
+            {
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine("Error: The current working directory does not contain the 'wwwroot' folder. Please run this application from the correct directory.");
+                Console.ResetColor();
+                return;
+            }
+
             try
             {
                 // this will override the config.json values if the relevant command line flags are present